Builds with refactor of first round of functions
ctors, and populate_element_candidates pulled into base class
