+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_sources(GlobalSourceList
    ABLProfileFunction.C
    ActuatorLinePointDrag.C
+   Actuator.C
    Algorithm.C
    AlgorithmDriver.C
    AssembleContinuityEdgeOpenSolverAlgorithm.C
